Factored out compiler step
