+++ cli/build.js
@@ -1,13 +1,13 @@
 /*jshint node: true*/
 'use strict';
 
-const logger = require('winston');
 const fs = require('fs-extra');
 const merge = require('merge');
 const skyPagesConfigUtil = require('../config/sky-pages/sky-pages.config');
 const generator = require('../lib/sky-pages-module-generator');
 const assetsConfig = require('../lib/assets-configuration');
 const pluginFileProcessor = require('../lib/plugin-file-processor');
+const runCompiler = require('./utils/run-compiler');
 
 function writeTSConfig() {
   var config = {
@@ -128,38 +128,12 @@ function build(argv, skyPagesConfig, webpack) {
 
   assetsConfig.setSkyAssetsLoaderUrl(config, skyPagesConfig, assetsBaseUrl);
 
-  const compiler = webpack(config);
-
-  return new Promise((resolve, reject) => {
-    compiler.run((err, stats) => {
-      if (err) {
-        logger.error(err);
-        reject(err);
-        return;
-      }
-
-      const jsonStats = stats.toJson();
-
-      if (jsonStats.errors.length) {
-        logger.error(jsonStats.errors);
-      }
-
-      if (jsonStats.warnings.length) {
-        logger.warn(jsonStats.warnings);
-      }
-
-      logger.info(stats.toString({
-        chunks: false,
-        colors: false
-      }));
-
+  return runCompiler(webpack, config)
+    .then(() => {
       if (compileModeIsAoT) {
         cleanupAot();
       }
-
-      resolve(stats);
     });
-  });
 }
 
 module.exports = build;
