+++ cli/e2e.js
@@ -188,10 +188,10 @@ function spawnServer() {
 /**
  * Spawns the build process.  Captures the config used.
  */
-function spawnBuild(argv, skyPagesConfig, webpack) {
+function spawnBuild(argv, skyPagesConfig) {
   return new Promise((resolve, reject) => {
     logger.info('Running build...');
-    build(argv, skyPagesConfig, webpack)
+    build(argv, skyPagesConfig)
       .then(stats => {
         logger.info('Build complete.');
         resolve(stats.toJson().chunks);
@@ -205,13 +205,13 @@ function spawnBuild(argv, skyPagesConfig, webpack) {
  * Assumes build was ran.
  * @name e2e
  */
-function e2e(argv, skyPagesConfig, webpack) {
+function e2e(argv, skyPagesConfig) {
   start = new Date().getTime();
   process.on('SIGINT', killServers);
 
   Promise
     .all([
-      spawnBuild(argv, skyPagesConfig, webpack),
+      spawnBuild(argv, skyPagesConfig),
       spawnServer(),
       spawnSelenium()
     ])
